[ZENTORCH CORE] Migrate WOQ linear implementation to stable ABI (#558)
* [ZENTORCH CORE] Migrate WOQ linear implementation to stable ABI

Switch WOQ linear, fusion, and repack kernels from at::Tensor to torch::stable::Tensor
Register WOQ ops via STABLE_TORCH_LIBRARY instead of TORCH_LIBRARY
Add stable tensor helpers in MatmulUtils.hpp for matmul/linear shape, contiguity, and post-op setup
Add stable set_zendnnl_tensor_attributes overload in Memory.hpp
Update AOTI shim entry points to use stable_from_handle and optional_stable_from_handle

+++ src/cpu/python/zentorch/_meta_registrations.py
@@ -814,10 +814,12 @@ def meta_zentorch_fused_moe(
 
 @register_meta("zentorch_woq_repack_weight")
 def meta_zentorch_woq_repack_weight(unpacked_weight):
-    # Returns a packed weight tensor of shape [N, K/8]
+    # Returns int32 packed weights of shape [N, K/8] (8 int4 values per int32).
     K = unpacked_weight.size(1)
     K_packed = K // 8
-    return unpacked_weight.new_empty((unpacked_weight.size(0), K_packed))
+    return unpacked_weight.new_empty(
+        (unpacked_weight.size(0), K_packed), dtype=torch.int32
+    )
 
 
 @register_meta("zentorch_woq_repack_from_int4pack")
